requirements.txt
# === Core API Framework ===
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.1.0

# === HTTP Client ===
httpx>=0.28.0

# === Model Gateway (LLM Abstraction) ===
litellm>=1.60.0

# === Vector Database ===
qdrant-client>=1.13.0

# === Memory Layer ===
mem0ai>=0.1.60

# === Embedding support (for local embeddings via sentence-transformers) ===
sentence-transformers>=3.4.0

# === PDF Parsing ===
pymupdf>=1.25.0

# === Testing ===
pytest>=8.0.0
pytest-asyncio>=0.24.0

# === Agents / Memory ===
ollama>=0.6.1

# === Podcast Agent ===
duckduckgo-search>=7.0.0
edge-tts>=6.1.0
beautifulsoup4>=4.12.0

# === Podcast Agent (Optional - Premium TTS) ===
# elevenlabs>=1.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# === Background Tasks (Phase 3) ===
arq==0.25.0
redis>=5.0.0

# === Messaging (Phase 4) ===
python-telegram-bot>=20.0
httpx>=0.28.0

# === Sandboxing (Phase 8, Optional) ===
# docker>=7.0.0
